+++ katelaskenta/scripts.katelaskenta.js
@@ -9,20 +9,20 @@ $(document).ready(function () {
 
     // Esitell‰‰n muuttujat.
     // Kaikki muuttujat alustetaan funktiossa myˆhemm‰ss‰
-    // vaiheessa. 
+    // vaiheessa.
     var tuoterivitTaulukko; // koko taulukko
     var tuoterivit; // kaikki taulukon rivit
     var tuoterivitCheckboxes; // tuoterivien checkboxit
     var footerRivi; // taulukon footer rivi
     var footerCheckbox; // taulukon footer checkbox
     var footerLaskeKaikki;  // Footer osion
-    
+
     // Kaikki sarake-p‰‰tteiset muuttujat ovat jqueryn
     // selectoreita kertomaan, miss‰ oikea sarake on
     // mik‰li k‰yttˆliittym‰‰ menn‰‰n muuttamaan.
-    var footerKateMyyntihintaSarake; 
-    var footerKateMyymalahintaSarake; 
-    var footerKateNettohintaSarake; 
+    var footerKateMyyntihintaSarake;
+    var footerKateMyymalahintaSarake;
+    var footerKateNettohintaSarake;
     var tuoteriviCheckboxSarake;
     var tuoteriviLaskeNappiSarake;
     var kateMyyntihintaSarake;
@@ -31,8 +31,8 @@ $(document).ready(function () {
     var myyntihintaSarake;
     var myymalahintaSarake;
     var nettohintaSarake;
-    
-    
+
+
     // Esitell‰‰n funktiot.
     // Toteutukset lˆytyv‰t alapuolelta.
     var alustaMuuttujat;
@@ -40,7 +40,7 @@ $(document).ready(function () {
     var onkoTyhja;
     var lisaaHintaanKate;
     var asetaUusiHinta;
-    
+
     /**
      * Funtion toiminto on vain alustaa tavittavat muuttujat, joita
      * eri toiminnallisuuksissa k‰ytet‰‰n. N‰in elementtien hakuja
@@ -53,7 +53,7 @@ $(document).ready(function () {
         footerRivi = tuoterivitTaulukko.find("tfoot tr");
         footerCheckbox = tuoterivitTaulukko.find("tfoot tr td:first-child input[type=checkbox]");
         footerLaskeKaikki = tuoterivitTaulukko.find("tfoot tr td:last-child a");
-        
+
         footerKateMyyntihintaSarake = "td:nth-child(4) input";
         footerKateMyymalahintaSarake = "td:nth-child(6) input";
         footerKateNettohintaSarake = "td:nth-child(8) input";
@@ -66,8 +66,8 @@ $(document).ready(function () {
         tuoteriviCheckboxSarake = "td:nth-child(2) input[type=checkbox]";
         tuoteriviLaskeNappiSarake = "td:last-child a";
     }
-    
-    
+
+
     /**
      * Funktio tarkistaa annetun myyntikatteen, jotta laskutoimitukset
      * voidaan suorittaa.
@@ -75,13 +75,13 @@ $(document).ready(function () {
      * Palauttaa false, jos virhe lˆytyy.
      */
     var onkoVirheellinenMyyntikate = function(myyntikate) {
-        if (isNaN(myyntikate)) 
+        if (isNaN(myyntikate))
             return false;
-        if (myyntikate >= 100 || myyntikate < 0) 
+        if (myyntikate >= 100 || myyntikate < 0)
             return false;
         return true;
     };
-    
+
     /**
      * Funktio tarkistaa onko annettu arvo tyhja.
      */
@@ -90,7 +90,7 @@ $(document).ready(function () {
             return true;
         return false;
     };
-    
+
     /**
      * Funktio lis‰‰ annettuun hintaan annetun katteen.
      *
@@ -102,16 +102,16 @@ $(document).ready(function () {
     var lisaaHintaanKate = function(keskihankintahinta, myyntikate) {
         var floatKeskihankintaHinta = parseFloat(keskihankintahinta);
         var floatMyyntikate = parseFloat(myyntikate);
-        
+
         if (onkoTyhja(floatMyyntikate)) {
             alert("Katekentt‰ ei voi olla tyhj‰.");
             return false;
         }
-        if (isNaN(floatKeskihankintaHinta)) {
-            alert("Virheellinen keskihankintahinta.");
+        if (isNaN(floatKeskihankintaHinta) || floatKeskihankintaHinta == 0) {
+            //alert("Virheellinen keskihankintahinta.");
             return false;
         }
-        
+
         if(!onkoVirheellinenMyyntikate(floatMyyntikate)) {
             alert("Virheellinen kate. Katteen pit‰‰ olla 0-100 v‰lill‰.");
             return false;
@@ -119,7 +119,7 @@ $(document).ready(function () {
 
         return floatKeskihankintaHinta / (1 - (floatMyyntikate / 100));
     };
-    
+
     /**
      * Funktio asettaa elementtiin uuden hinnan.
      *
@@ -134,17 +134,17 @@ $(document).ready(function () {
                                 .text(hinta.toFixed(2));
         $(kohdeElementti).empty().html(htmlUusiHinta);
     };
-    
+
     /**
      * T‰st‰ l‰htien ohjelmakoodissa m‰‰ritell‰‰n elementeille niiden
      * toimintalogiikka aikaisemmin esitettyjen funktioiden avulla.
      * Kaikkien funktioiden ja yleisten muuttujien kuuluisi olla esitetty
      * ennen seuraavia toimenpiteit‰.
      */
-    
+
     // Kutsutaan muuttujien alustus.
     alustaMuuttujat();
-    
+
     // Lis‰t‰‰n jokaiselle tuoterivin laske-painikkeellle toimintalogiikka.
     // Laske painike laskee annetun kateprosentin mukaan uuden hinnan.
     $.each(tuoterivit, function () {
@@ -152,33 +152,33 @@ $(document).ready(function () {
 
         var myyntikate = $(this).find(kateMyyntihintaSarake);
         var myyntihintaElementti = $(this).find(myyntihintaSarake);
-        
+
         var myymalakate = $(this).find(kateMyymalahintaSarake);
         var myymalahintaElementti = $(this).find(myymalahintaSarake);
-        
+
         var nettokate = $(this).find(kateNettohintaSarake);
         var nettohintaElementti = $(this).find(nettohintaSarake);
-        
+
         $(this).find(tuoteriviLaskeNappiSarake).on("click", function (event) {
             event.preventDefault();
             var uusiMyyntihinta = lisaaHintaanKate(keskihankintahinta, myyntikate.val());
             var uusiMyymalahinta = lisaaHintaanKate(keskihankintahinta, myymalakate.val());
             var uusiNettohinta = lisaaHintaanKate(keskihankintahinta, nettokate.val());
-            
+
             if(uusiMyyntihinta !== false && myyntikate.val() > 0) {
                 asetaUusiHinta(uusiMyyntihinta, myyntihintaElementti);
             }
-            
+
             if(uusiMyymalahinta !== false && myymalakate.val() > 0) {
                 asetaUusiHinta(uusiMyymalahinta, myymalahintaElementti);
             }
-            
+
             if(uusiNettohinta !== false && nettokate.val() > 0) {
                 asetaUusiHinta(uusiNettohinta, nettohintaElementti);
             }
         });
     });
-    
+
     // Lis‰t‰‰n taulukon viimeisen rivin valintaruudulle toimintalogiikka.
     // Ruutua klikkaamalla joko valitaan kaikki tai poistetaan valinta
     // kaikista ruuduista.
@@ -193,24 +193,24 @@ $(document).ready(function () {
                 });
             }
         });
-    
+
     // Lis‰t‰‰n taulukon viimeisen rivin "laske kaikki" -painikkeelle
     // toimintalogiikka. Painiketta painaessa lasketaan uusi hinta ja
     // samat arvot m‰‰ritet‰‰n taulukon kaikille muille tuoteriveille.
      footerLaskeKaikki.on("click", function (event) {
         event.preventDefault();
-        
+
         var myyntikate = tuoterivitTaulukko.find("tfoot tr").find(footerKateMyyntihintaSarake).val();
         var myymalakate = tuoterivitTaulukko.find("tfoot tr").find(footerKateMyymalahintaSarake).val();
         var nettokate = tuoterivitTaulukko.find("tfoot tr").find(footerKateNettohintaSarake).val();
-        
+
         if (!onkoTyhja(myyntikate)) {
             if(!onkoVirheellinenMyyntikate(myyntikate)) {
                 alert("Virheellinen kate. Myyntikatekentt‰ ei voi olla tyhj‰ ja katteen pit‰‰ olla 0-100 v‰lill‰.");
                 return true;
-            }    
+            }
         }
-        
+
         if (!onkoTyhja(myymalakate)) {
             if(!onkoVirheellinenMyyntikate(myymalakate)) {
                 alert("Virheellinen kate. Myymalaatekentt‰ ei voi olla tyhj‰ ja katteen pit‰‰ olla 0-100 v‰lill‰.");
@@ -224,7 +224,7 @@ $(document).ready(function () {
                 return true;
             }
         }
-        
+
         // K‰yd‰‰n jokainen rivi l‰pi ja asetetaan uusi hinta, jos hinta
         // ei ole virheellinen.
         $.each(tuoterivit, function () {
@@ -241,7 +241,7 @@ $(document).ready(function () {
                     }
                     $(this).find(kateMyyntihintaSarake).val(myyntikate);
                 }
-                
+
                 if (!onkoTyhja(myymalakate)) {
                     var myymalahintaElementti = $(this).find(myymalahintaSarake);
                     var uusiMyymalahinta = lisaaHintaanKate(keskihankintahinta, myymalakate);
@@ -250,7 +250,7 @@ $(document).ready(function () {
                     }
                     $(this).find(kateMyymalahintaSarake).val(myymalakate);
                 }
-                
+
                 if (!onkoTyhja(nettokate)) {
                     var nettohintaElementti = $(this).find(nettohintaSarake);
                     var uusiNettohinta = lisaaHintaanKate(keskihankintahinta, nettokate);
@@ -262,5 +262,5 @@ $(document).ready(function () {
             }
         });
     });
-    
+
 });
